+++ packages/agent/CHANGELOG.md
@@ -1,5 +1,7 @@
 # Changelog
 
+## [Unreleased]
+
 ## [0.74.57] - 2026-05-29
 
 ### Fixed